Create scripts folder
