+++ .env.sample
@@ -1,4 +1,5 @@
 NODE_ENV=
+ALGOLIA_INDEXING=true
 ALGOLIA_APPLICATION_ID=
 ALGOLIA_API_KEY=
 ALGOLIA_INDEX=
