+++ tsconfig.json
@@ -12,7 +12,8 @@
         "strictNullChecks": false
     },
     "include": [
-        "src/**/*"
+        "src/**/*",
+        "scripts/*"
     ],
     "exclude": [
         "node_modules"
